feat(town): extract castle keep placement to layout_core.js and wire town_gen to use placeCastleKeep; import getCastleKeepSizeConfig.
